Add function to get current courses.
The API returns courses with old fromData and toDate. Provide a convenience function that only returns courses between those dates.

I'm not sure under what other circumstances the API will return more than one course. If the time range is the only factor, then it may be useful to return a dictionary rather than a list of a single dictionary.